added ability to link store product without submitting item form. Can only happen if product already created, meaning in updateview

+++ newco/apps/items/views.py
@@ -67,16 +67,18 @@ class ContentFormMixin(object):
         form = self.load_form(request)
         if "next" in request.POST:
             self.success_url = request.POST.get("next")
-        if self.model == Item and "store_search" in request.POST:
-            form.stores_search()
-            return self.render_to_response(self.get_context_data(form=form))
-        elif self.model == Item and "remove_links" in request.POST:
-            if self.object:
+        if self.model == Item and ("store_search" in request.POST or
+                                    "add_links" in request.POST or
+                                    "remove_links" in request.POST):
+            if "add_links" in request.POST and self.object:
+                form.link_aff(self.object)
+            if "remove_links" in request.POST and self.object:
                 aff_item_ids = request.POST.getlist("linked_aff")
                 linked_items = self.object.affiliationitem_set.select_related()
                 aff_items_to_delete = linked_items.exclude(id__in=aff_item_ids)
                 for aff_item in aff_items_to_delete:
                     aff_item.delete()
+            form.stores_search()
             return self.render_to_response(self.get_context_data(form=form))
         elif form.is_valid():
             return self.form_valid(form)
